requirements.txt
pelican
Markdown
typogrify
awscli
python-dotenv
